# sdk/python/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_ingest_file_sync_api_v1_ingest_file_post": {
        "properties": {
          "agent_id": {
            "default": "default",
            "title": "Agent Id",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_ingest_file_sync_api_v1_ingest_file_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestSyncResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/SlimEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "edges_created": {
            "title": "Edges Created",
            "type": "integer"
          },
          "edges_merged": {
            "title": "Edges Merged",
            "type": "integer"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/SlimNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "nodes_created": {
            "title": "Nodes Created",
            "type": "integer"
          },
          "nodes_merged": {
            "title": "Nodes Merged",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "agent_id",
          "nodes_created",
          "nodes_merged",
          "edges_created",
          "edges_merged",
          "nodes",
          "edges",
          "latency_ms"
        ],
        "title": "IngestSyncResponse",
        "type": "object"
      },
...
      "SlimEdge": {
        "description": "Edge without redundant fields.",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "relation"
        ],
        "title": "SlimEdge",
        "type": "object"
      },
      "SlimNode": {
        "description": "Node without embedding \u2014 used in API responses to avoid sending 1536 floats.",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "properties": {
            "default": {},
            "title": "Properties",
            "type": "object"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "type"
        ],
        "title": "SlimNode",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SpAIder \u2014 A multi-agent semantic knowledge graph platform powered by LLMs and Neo4j. Ingest unstructured text, build a persistent knowledge graph, query it in natural language, and synthesise fine-tuning datasets for downstream models.",
    "title": "SpAIder",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingest/file": {
      "post": {
        "description": "Thin wrapper around the Upload Connector. Accepts any supported document format \u2014 PDF, DOCX, PPTX, HTML, Markdown, and plain text \u2014 via `multipart/form-data`.\n\nInternally delegates to `UploadConnector` \u2192 `parser_service.parse()` \u2192 full extraction pipeline. The response shape is identical to `/ingest/sync` for backward compatibility.",
        "operationId": "ingest_file_sync_api_v1_ingest_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_file_sync_api_v1_ingest_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestSyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a single file for synchronous ingestion (PDF, DOCX, PPTX, HTML, MD, TXT)",
        "tags": [
          "ingest"
        ]
      }
    },
...
  }
}